audio/mpq8064: Miscellaneous fixes for the HAL
Calling the setparameter interface from broadcast stream
handle gives a crash.
Broadcast stream handle was typecasted to the
audio stream handle resulting in a crash. Modified to the
proper typecasting.
Removing the device from the device list after the close of device.
Making the device handle to NULL after removing form the list.

Change-Id: I468a29925ce7fd73cf94165b9553f9f569f57a13