audio/msm7627a: Fix for FM mute during touch tone playback
-FM mutes when touch tone playback happens when an application
is selected.
-CheckDeviceMuteStrategies checks for popcount more than or equal
to 2 to mute a strategy. If DEVICE_OUT_FM is present in the devices,
this should be 3 instead of 2.
-Extend CheckDeviceMuteStrategies and getDeviceForVolume from
base class to check popcount against 3 instead of 2 if FM device is
selected. Temporary mute during device switch should happen only when
getDeviceForVolume for both the devices is different and not just if
both the devices are different. Also add check for FM device to return
DEVICE_OUT_FM in getDeviceForVolume.

CRs-Fixed: 394066
Change-Id: Icf02f2ed72d60e6747db56bf4ef520f04f60cbfe