audio/msm7627a: Fix for native crash in LPA playback
-If seek and next operation is performed during LPA playback,
native crash is observed sometimes.
-Avoid lock in flush to allow flush call to complete during write.
Protect empty and filled queues by adding appropriate lock to avoid
race conditions with write and write done event.
Also if no data to write to driver, return immediately from write so
that the buffer is not pushed to filled queue and corrupt it.

Change-Id: Ie64942ea8f82955ef92c73cad05dbba4f6b46d85
CRs-Fixed: 395477