audio/msm7627a: Fix for seek and pause timeout issues in LPA
-Audio is distorted after doing seek during LPA playback.
A buffer is held with audio HAL just before flush and it is sent
to driver immediately after flush. This results in previous buffer
to be played after seek. Set mSkipWrite flag just before flush so
that the write call skips any buffer held with HAL during flush.

-After Pausetimeout, reset is called due to which EOS is posted
to observer. This results in start or reset operation on a closed
sink leading to crashes. Avoid posting EOS if reset is already
called.

Change-Id: If6e30d933c5eeaa9f958041e4977b1867079f097