alsa_sound: Fix for no audio in BT after standby is called.
When BT is connected, and there is no audio for three seconds,
standby is called. But pcm device close was bypassed in the case,
when BT device was connected. Due to this mRouteAudioToA2dp was not
getting set to true in subsequent playback and hence a2DpPlayback
was never getting started.

Make a fix to close pcm device even when BT device is connected.

(cherry picked from commit 41d8544b953cc5f0f3d8ac0316df60ff631811fc)

Change-Id: I21aa6942319a97baeb7ee9ee6507cb679cf9c001
CRs-Fixed: 409889