alsa_sound: Fix for no audio in BT after standby is called.
When BT is connected, and there is no audio for three seconds,
standby is called. But pcm device close was bypassed in the case,
when BT device was connected. Due to this mRouteAudioToA2dp was not
getting set to true in subsequent playback and hence a2DpPlayback
was never getting started.

Make a fix to close pcm device even when BT device is connected.

Change-Id: Ice455455063179c82640b9fef9fdab7c002b615c
CRs-Fixed: 409889