audio: mpq8064: Fix Rx playback fail during MT call
Add support for VOIP RX profile in audio policy config file and
output stream open.
In the use case of the voip- voice over gtalk/SIP, the caller tunes
and messenger start tones are played only on Speaker even though
the device is configured to HDMI and Speaker. Add support for the
same.

CRs-Fixed: 402243
Change-Id: I919a73cf4f0195286a49dc67328cf6886aa5d713