audio/mpq8064: fix for VOIP over BT SCO.
-Add a new variable to keep track of the old mode of the voice
call, if there is a change in the mode the device routing has to be done.
-Update the UCM config file to solve the issue when the HAL is trying to
open a specificed usecase and if the use case is not availabe UCM will fall
back to the default device and opens SLIM bus backend,
and this turns HAL to waits for the buffer from two backends.
-Removed the check from the device switch when the old device and
current device are same, this will effect when themode is changed the device
routing is not hanppening properly and this is how 8960 HAL implemented.

CRs-Fixed: 405892
Change-Id: I6cc424cecd1d48586baaa757f370e8271819ccd0