audio/mpq8064: Update openInputStream to take mcurdevice
For the voip use case the devices from the device strategy
Will be DEVICE_IN_COMMUNICATION, but we need to route the device
To the current enabled device.

Update wired headset as input device for the DIGITAL_OUT_PROXY in
getDevice logic as we do not have the DMIC device on MPQ target.

CRs-Fixed: 402243
Change-Id: I0926348a60188f5e995b14711c981f29289a46c4