audio/msm7627a: Fix for wrong call volume in MO call when FM is on
- With FM in back ground, when MO call is made, the actual voice call
volume doesn't match with the voice volume set by the user.
- When we change FM volume, the same volume level is applied on Voice
call stream also. At the beginning of MO call, a check is done to see
if the current voice volume is same as the previous value. If yes,
voice call volume is not set. But, FM is changing stream volume from
the last voice call value, which is not reflected as the last set
voice volume.
- Update the last set voice volume to reflect change due to FM volume
setting.

Change-Id: Ie38743ae3da6e4d2baaf96914653d1b697f16551
CRs-Fixed: 417830