audio/mpq8064: AC3 compressed playback stuck for repeat mode on
During the stop call to HAL it acquires a lock to close down the devices
and the event thread.
Also for the eventThread we need a lock to process the events,
so this may result in a deadlock.
Try acquiring the lock in eventThread and if it fails check for the
flag mKillEventThread to kill the thread
this will ensure that no deadlock has occured.

Change-Id: I7815a5ff6cf5686bd0b4dfcb8d4d9cb2989a3702