audio/mpq8064: AC3 compressed playback stuck for repeat mode on
During the stop call to HAL it acquires a lock to close down the devices
and the event thread.
Also for the eventThread we need a lock to process the events,
so this may result in a deadlock.
Try acquiring the lock in eventThread and if it fails check for the
flag mKillEventThread to kill the thread
this will ensure that no deadlock has occured.

CRs-Fixed: 426071

(cherry picked from commit b7e0aa891d8136679c685b21905b9b47109be644)

Change-Id: I27caf5406534c5c3a40b66811b428bce5c09e668