alsa_sound: Issue with seek and write
-Song repeat was heard when seeked in short steps
- If seek is issued when a write is going on, we have
to skip the buffer. Also, we have to clear the skip flag
if there is data coming back from the driver
- Check the flag set from the LPAPlayerALSA level to check
if the flag is got after seeking
- Also, make sure pause is not called only if write was called
- And that buffers have to be flushed as many times as
when seek is called. For this, a count is introduced

(cherry picked from commit 86370a871b80bcfa5ae721758f5eb1c701789d36)

Change-Id: I1e8709d1b319ef37feb92930c4723e9b9ed3773a
Crs-Fixed: 428750