audio/mpq8064: Fix for the stuck at the end of the playback and choppy output
- Playback hangs at the end of the playback for few DTSHD clips.
The sync_ptr is called from the pcm_write and in the event thread.
Concurrent calls can cause the incorrect update of the hw_ptr & appl_ptr

Synchronising the calls for the sync_ptr to avoid the race conditions
in the update of the hw_ptr & appl_ptr.

- For any write done events sync the hw_ptr before updating the buffer queues.

- The transfer of the buffer from the filled queue to empty queue should be gaurded against
  any modification in either of the queues

CRs-Fixed: 427390 433021
(cherry picked from commit 740edbce195926cfb5139da96c1bed2a334cd403)

Change-Id: I194c5bc0ba5f961d7b5290b02cb7b2902474f95c