audio/mpq8064: Media server crash for Vcap PCM input
Issue: Invalid memroy access while reading capture data is
due to the overflow of an local variable. Also an expression
involving two variables which are int_32 results in a uint_64
value but as both the variables are uint32 the result truncates
to unit_32, this causes wrong offset to be calculated.

Fix: Modified the local vairable to be uint_64 ensuring that it does
not overflow. Also typecast one of the unit_32 as unit64 to enusre
the result is not truncated.

(cherry picked from commit 5cb8f7b9fdbf850192799c705f3fda76ad494e27)

Change-Id: I7dff7fe232ad9b603c52a24fd5b9c53f94af1cb0