audio/mpq8064: Fix the WFD BT routing concurrency
WFD queries the getParameters for the device list,
Since the mCurDevices in AudioHardware doesn't popultate the
devices present in the AudioSessionOut it may return the incorrect devices.

Iterate over the device list and return the all the device present.

For the WFD use case the audio shouldn't be routed to SPDIF,
Removing the SPDIF from the device list when the output device is
proxy and the BT is disabled.

CRs-Fixed: 439194
Change-Id: I083c75b2f1e48f24d081884967481e7007b64eb8