audio/msm7627a: Set phone device only for enforced audible
-Wrong device gets updated in strategy cache for strategy media.
-Currently Strategy_Phone device is selected for enforced audible
if not enforced and also for strategy media. Due to this phone
device gets updated in cache for strategy media. The expected
behavior is not only that strategy phone should get higher priority
to strategy media in getNewDevice() but also update correct device
in the cache for each strategy. Due to the current implementation,
this behavior is changed.
-Set to strategy phone device only in cases when strategy is
Enforced_Audible and not enforced.

Change-Id: Iac67bf3b00f3d298349f156fc177d58f14ec0b9f
CRs-Fixed: 440114