audio/mpq8064: Fix crash for device switch from HDMI to BT
Issue:
Invalid ALSA handle is accessed when we switch from HDMI to
A2DP.In doRouting for A2DP devices we resume the playback even if the
playback was paused from the framework.
Fix:
Resume playback in doRouting only if the playback was paused in doRouting,
otherwise the stream is resumed by the framework.
Remove forced start as that is never executed.

CRs-Fixed: 441160

Change-Id: Iaa5c354940b40568393019ea28ab91fd2f31dad4