audio/mpq8064: Fix infinte wait in HAL in case of error scenarios
Issue: Eventhread on start waits on a mutex, which is signald from
the write api or reset api, in some error case if directly the reset
api is called and it gets scheduled before the evethtread getting
scheduled even once then it would miss the signal from reset(), and
wait on the mutex forever.
Fix:Ensure that if reset comes first then we dont wait on the mutex
after signal is sent.

Change-Id: Ibbf48bcf519ae5deb98ceeb142c4409476606498
CRs-Fixed: 442542