libalsa-intf: Add support to handle numerical amix commands
The numerical amix commands from UCM files are not supported in
the current code unless they are the amix commands to set volume.
But there are other cases such as sample rate, which are defined
as numerical commands as well.  The change is to add the support
to cover rest of the use cases.

The change includes update to amix commands for BT SCO sample
rate from string type to numerical type.

(cherry picked from commit e52afb69d607814a58398a65bcaf7599ac9b95dc)

Change-Id: I3b375ee95570219d65b024069f41f6a7763105d1
CRs-fixed: 394165