libalsa-intf: Add support to handle numerical amix commands
The numerical amix commands from UCM files are not supported in
the current code unless they are the amix commands to set volume.
But there are other cases such as sample rate, which are defined
as numerical commands as well.  The change is to add the support
to cover rest of the use cases.

The change includes update to amix commands for BT SCO sample
rate from string type to numerical type.

Change-Id: I0ca2b8e836d0dea96b6f5cb9a34b907e5990b295
CRs-fixed: 394165
(cherry picked from commit e52afb69d607814a58398a65bcaf7599ac9b95dc)