audio/mpq8064: Fix the playback hangs after multiple operations
On performing multiple operations on the playback of the dts clips
the playback is stuck in the middle.
On the seek operation the appl_ptr is updated in the pcm_prepare call
which is overwritten by the eventThread.

Fix is to sync the appl_ptr from the kernel each time.
Added the flag SNDRV_PCM_SYNC_PTR_APPL during the sync_ptr call.

Change-Id: Iae420da272165566de1e8e9b0008158a933d4d57
CRs-Fixed: 436946