audio/mpq8064: Fix for no Audio on LG TV when used as hdmi sink
Issue: We configure the output number of channels based on the EDID data of the
connected HDMI sink device.
-In case of LG TV, it supports two formats, AC3 and LPCM with 6 and 2
number of channels respectively.
-The channel map provided in the EDID data is always for uncompressed
audio data i.e.LPCM format (CEA-861-D Version 3),which is
FL/FR in case of the LG TV as the number of output channels for LPCM is 2.
- Current implementaion relies on the maximum number of channels
supported by the sink irrespective of the audio format, because of this we
configure HDMI out for 6 channels, and the channel map(channel allocation)
as 0 for FL/FR.
- As there is a mismatch between the channel allocation and the number of output
channels no audio is heard.
Fix: Configure the number of output channels only with the max number of channels
supported for LPCM format based on the hdmi sink EDID data.

Change-Id: I3b43d1be4224d951917c466184d617181d16490e
CRs-Fixed: 452228