audio/mpq8064: Volume applied on start not effective for AAC clips.
- For AAC/AC3 clips, we check fro sample rate after every frame decode,
if there is change in the sample rate from what was set earlier, we
close the driver and reopen it with the new sample rate.
- The volume which is sent for the earlier instance of the driver
is not sent again as the framework is not aware of the clsoing and opening
of the driver, and hence the new driver opens with default volume.
- We have the volume applied on the earlier instance of the driver, now we
send that volume on the new istance of the driver as soon as it is open.

Change-Id: Ibd84e8f68ed1ac6e43ff4a48bbbea7764859b4fd