alsa_sound: Fix to enable wired headset mic during VoIP call.
-When in gtalk voice call with wired headset, voice call is routed
to ear piece in stead of wired headset.

-In gtalk voice call starts over wired headset, in setParameters(),
by default DEVICE_BUILTIN_MIC was getting selected, because of which
Rx routes to ear piece instead of headset.

-Appropriate input (Tx)/output (Rx) device is now selected based on
the device connected. Also there was a bug in parsing KeyValuePair
in HAL, because of which by default DEVICE_BUILTIN_MIC was getting
selected.

CRs-Fixed: 483752

Change-Id: I56d59c53ab1987b185147156fb11757c28c298d1