audio/mpq8064: Synchronize enable and disable sequence across streams
Issue:- Current sequence is enable device, open device handle and then
add the device handle to device list.
The stream which is closing refers to this device list to confirm
if no other active stream is using the same device, if not then the
device is disabled.
If between the enable device and addition to device list, disable
device is invoked, then the device is disabled but the other stream
does not enable the device and it results in a session with
no active device.
Fix:
  For BroadCast session guard the device routing with a lock which
is shared across all active sessions.

CRs-Fixed: 492447
Change-Id: Ib5deea161f53fdcb31b1f635889d05cb49bbb4e8