audio/mpq8064: Fix the target crash with the 24 bit playback
24 bit playback requires the 1 auxillary byte of data with each sample.
This addition of byte over the complete data was corrupting the meta dat.

Fixed by adding the auxillary bytes only to the valid data.
Decrease the time per period to 20ms to keep the period sizes in check
for the large sample rates. Also it will reduce the latency to fill the buffers
by the application.

Change-Id: I84558b0c12345e5cfc2cc3c20d0e52cb364cb4ad