audio/mpq8064: Lower the priority of the SPDIF device in PolicyManger
As SPDIF device is added by default in the session,
lower the priority of the SPDIF device which will be selected
if no other active device is present.

DEVICE_OUT_USB_DEVICE is an alias for the SPDIF device.
This is due to reason that Android doesn't support the SPDIF
as a device and all the 32 bits are used up.

CRs-Fixed: 524076
Change-Id: I7abdfa130bbc65eb607cc34d93670cd12a79d45d