audio/mpq8064: Fix the missing and the half touch tone issue
Issue: Due to the external sink dropping the data during the
intialization sequence, the intial audio data is dropped for
an audio session.
Fix: Run a session always to ensure there is no drop of data
during the setup. If no data is rendered the AFE will insert the
silence data packet on the interface.
To ensure the system goes into the suspend, close the session when
audio HAL receives the screen_state=on event and resume on the receipt
of the event screen_state=off
Introduce a new identifier _disverb, to facilitate the deroute of
specific devices from the verb.

CRs-Fixed: 495470
Change-Id: Ib6fcb45ab002c1acd891127fa63ee3e0733ee6dc