audio: Add support for multiple VoIP sessions
Two VoIP sessions will be created, one is for
pcm format to support VoIP or Qchat call,
another is for compressed audio format to support
VoWLAN and Qchat.
In order to make sure VoWLAN is not fail during VoIP
and VoWLAN concurrency, VoWLAN call will use the session
with compressed audio format.

Change-Id: Icee621e2832c23777013393ad98cc69e54fe0d01