alsa_sound: route STREAM_BT_SCO to DEVICE_BT_SCO in STRATEGY_PHONE
- getDeviceForStrategy() returns DEVICE_EARPIECE for STRATEGY_PHONE
when BT_SCO device is connected but before FORCE_BT_SCO is set. This
is causing the initial volume setting provided by the BT SCO headset
to be cached improperly and overwriting the volume for
[DEVICE_EARPIECE, STREAM_BT_SCO] instead of [DEVICE_BT_SCO, STREAM_BT_SCO].
- Later when voice call is connected on BT SCO device, the volume bar on
the UI shows incorrect setting due to the mismatch between the BT SCO
headset's volume and the cached volume in Audio for the BT_SCO device.
- Fixed by returning proper device for the BT_SCO stream and there by
caching the initial volume properly.

Change-Id: I9ec1882fac50fe805d374e1335d1210784d1bdd6
CRs-fixed: 553243