alsa_sound: cache call states if one or more calls are active
In multisim scenario, when a call is active on one sub and a
incoming call is received on the second sub, the audio mode
will be changed to AUDIO_MODE_RINGTONE. Call states are not
cached if the mode is not AUDIO_MODE_IN_CALL.

As a result, when telephony updates the first call state to LCH,
it is not cached as the  the mode is AUDIO_MODE_RINGTONE and the
call on first sub is not moved to LCH state. When the second call
is started, voice data will be merged between both subs.

Cache states by checking if one or more calls are active
instead of checking for audio mode.

Change-Id: I0be30f2eb15c6ef0877aab179cecb0a9ad9e9439
CRs-fixed: 556043