audio/mpq8064: Fix the random audio noise
Since the routing command comes with 0 when the stream is paused
there is no way to identify that the HDMI device has been disconnected.
If the AFE port is left open while the HDMI core is shutdown, it may lead to
random noise or other artifacts.
Deroute the HDMI device when the stream goes to pause state and the HAL recieves
a doRouting command

CRs-Fixed: 563834
Change-Id: I2acd63f453074ced545958f92d627da0095e699e