audio/mpq8064: Fix the channel allocation to 0 for the 2 channels
For the sink device which advertises the 8 channel allocation block
and only 2channel LPCM block. This results in the mismatch between the channel
allocation blo0ck and the number of channels.
Due to this audio artifacts are observed during the playback on the
HDMI  interface
Fix the issue by fixing the channel allocation to 0 for the 2 channels config.

Change-Id: I6ffb1df8c3a04ee32c9e1659d0ad522564dbdc52