audio/mpq8064: Fix AVsync issues due to HDMI hotplug
Disconnection from HDMI sink results in derouting of the
session attached to HDMI.
On HDMI sink is reconnection, new session is created,
due to this there is an offset in the data buffered in the other
already open session(SPDIF) and the new session.
Delay the new session path with the buffering in the already opened
session to maintain AVSync.

Change-Id: I46b25640f925e44f0475925e2dc1c825a394e202