hal: Send voice cal data before routing voice usecases
-The sound intensity is drastically reduced in a DSDA call when
a particular device is re-selected after a device switch.
It is observed that for the active sub, acdb data tables are
registered to ADSP even before the right acdb tables are
sent to the acdb driver. Therefore, the tables which are infact
registered to the ADSP correspond to the previous device.
- Load the acdb files into the kernel memory before routing
voice usecases to prevent any incorrect tables being sent.

CRs-Fixed: 694397
Change-Id: Id2f2a43c887d9b9e74cd98858013d1ef655a93e3