hal: fix low volume issue during device switch in a hfp call
When an active hfp call is switched from the hands-free client
to the audio gateway and then back to the hands-free client,
voice is not heard at the client.

After the switch from the gateway to the client, the volume applied
is very low and hence is not audible.

Fix the issue by setting the volume properly.

CRs-Fixed: 720804
Change-Id: I4dfdff7d0b8e5763dc0a19d55531252994db90c3