audio_hal: increase written frames size when HAL is in SSR
Written PCM frames count is not increased when HAL stays in SSR,
it can make mismatch with the written framecount in AudioFlinger,
so rendered frames count return from AudioFlinger is not correct.
Then A/V sync will be missed since the played audio duration is
calculated by rendered frames count.

Increase the written frames count even the frames are not written
to the PCM driver during SSR.

CRs-Fixed: 791738

Change-Id: I53cfeae9294ae1f5a6466a749da90d1f8ef297cb