Revert "hal: Fix voice session setup failure"
Camera click tones are not playing while in voice calls and headset
is connected since get_output_devices doesnt return a valid device.
Fix by returning combo device speaker-and-headset when in voice call.

The original issue reported in CR 859796, seem to customer specific,
as on internal platforms, there are no usecases such as
"volte-call speaker-and-headset".

This reverts commit 62d96d50dafae78818994425275a743b68d70881.

Change-Id: Ic8146d8ed5ab7312fa40242c5fda00996a27c083