audio: Updated correct end point for Dolby
When check_usecases_codec_backend switches the device for all
usecases that share same codec backend, only out_snd_device
is set for that usecase but devices is still having the
old device. Wrong end point is passed to Dolby DS2 driver due
to this reason

Fix is to update the endpoint even when if routing is already done
for this usecase.

Change-Id: Ib2dd85a636bf840a22ea7116c5f84889d4218423